Fem: Move FemFrameSourceAlgorithm class to its own source files
