+++ src/Mod/Fem/App/VTKExtensions/CMakeLists.txt
@@ -1,0 +1,6 @@
+SET(FemVTK_SRCS
+   vtkFemFrameSourceAlgorithm.h
+   vtkFemFrameSourceAlgorithm.cpp
+    )
+
+target_sources(Fem PRIVATE ${FemVTK_SRCS})
